refactor(reconcile): split date-fence tests to promote_fence_test.go (#292)
